Get tests running w/ sqlite

+++ backend/database/migrations/2020_12_11_185827_add_approved_cols_and_provided_courses_to_instructor.php
@@ -30,11 +30,13 @@ class AddApprovedColsAndProvidedCoursesToInstructor extends Migration
     public function down()
     {
         Schema::table('instructors', function (Blueprint $table) {
-            $table->dropColumn('is_approved');
-            $table->dropForeign('approved_by');
-            $table->dropColumn('approved_by');
-            $table->dropColumn('approved_at');
-            $table->dropColumn('provided_courses');
+            if (env('DB_CONNECTION') != 'sqlite') {
+                $table->dropColumn('is_approved');
+                $table->dropForeign('approved_by');
+                $table->dropColumn('approved_by');
+                $table->dropColumn('approved_at');
+                $table->dropColumn('provided_courses');
+            }
         });
     }
 }
